Added support for tunneling PUT and DELETE over POST.
Added `enable_POST_tunneling` config directive. It looks for a GET or POST param named '$method' or a header named 'X-HTTP-Method-Override'. If it finds one of these with a valid value, the request method is changed to that value. Any and all of these items will be removed from the request before passing it along to the application.

There is a new implicit convention that request parameters used by pyramid_restler will start with a '$'.