Added collection filtering and ordering capability to SQLAlchemy ORM context.
Instead of always returning the entire collection for an entity, `SQLAlchemyORMContext.get_collection` now accepts various query-altering keyword args, most of which have the effect of filtering the query so that only a subset of the entire collection is returned:

- filters (a map of {filter name => filter value})
- distinct
- order_by
- limit
- offset

In general, the `get_collection` method for all context implementations may accept filtering and other query-altering keyword args.

In addition, a list of default `filters` can be specified at the class level. These filters will be applied first (to all queries) and then the args listed above will be applied (in the order listed).

TODO: Parse corresponding query parameters in view `get_collection` and pass them through to model `get_collection`.