Refactored `SQLAlchemyORMContext.to_json()` so that JSON object construction and wrapping can easily be customized.
Added `get_json_obj()` for constructing the object to be JSONified. Added `wrap_json_obj()` for wrapping that object.

Made one small, backward incompatible change: results are always returned in a list. Before, a single result would not be wrapped in a list. Always using a list makes result parsing on the client more consistent.