Pass per-request collection keyword args from view to context.
To set per-request keyword args, send a query param named $$ with a JSON object whose keys correspond to the context's `get_collection` keyword args.

Note: this is a first pass at implementing this. Not sure $$ is the best name. Also, this might be a generally useful pattern for sending action-specific keyword args via request params.